moved BlogRender into blogPage
